fix: divide frontend project into 3 sub app
