+++ app/app/components/wallet/WalletBar.tsx
