+++ mythos_reply/env.example
@@ -1,0 +1,18 @@
+# Database
+DATABASE_URL="postgresql://username:password@localhost:5432/mythos_reply"
+
+# Twitter API Credentials
+TWITTER_API_KEY=your_twitter_api_key
+TWITTER_API_SECRET=your_twitter_api_secret
+TWITTER_BEARER_TOKEN=your_twitter_bearer_token
+
+# NextAuth Configuration
+NEXTAUTH_SECRET=your_nextauth_secret_key
+NEXTAUTH_URL=http://localhost:3000
+
+# JWT Secret
+JWT_SECRET=your_jwt_secret_key
+
+# OpenRouter Configuration
+OPENROUTER_API_KEY=your_openrouter_api_key
+OPENROUTER_BASE_URL=https://openrouter.ai/api/v1
